# fix onnx export of speech foundation models