@ylacombe 2024-12-02
Thanks @nikosanto13, it's much clearer now. Looks like it's not a costly operation. This PR looks great to me.

cc @ArthurZucker  or @Rocketknight1 for a core maintainer review!